cn0577: Update xdc. Slack -1.4ns
Set a negative output delay in order to make checks on the
next clock period, otherwise it shows the Required Time at
Destination Clock Path as being negative.

In the Path Report for the path where timing fails, it shows
the Requirement as 0.225ns although it should be at least
one clock period (8.333ns).
For other paths, it shows the Requirement correctly.

Net delays vary depending on the output delay set to cnv_en (why?)